+++ CPPExamples/operators/include/FlattenLayer.h
@@ -10,7 +10,7 @@ public:
     FlattenLayer();
 
     // Forward pass
-    xt::xarray<double> forward(const xt::xarray<double>& input_data);
+    xt::xarray<float> forward(const xt::xarray<float>& input_data);
 
     // Get output shape
     std::vector<size_t> get_output_shape() const;
